+++ src/util/math.go
@@ -2,14 +2,14 @@ package util
 
 type MathUtil struct{}
 
-func (mu MathUtil) Max(a float32, b float32) float32 {
+func (mu MathUtil) Max(a float64, b float64) float64 {
 	if a > b {
 		return a
 	}
 	return b
 }
 
-func (mu MathUtil) Min(a float32, b float32) float32 {
+func (mu MathUtil) Min(a float64, b float64) float64 {
 	if a < b {
 		return a
 	}
